pyproject.toml
[project]
name = "ai-newstand"
version = "0.1.0"
description = "Daily newsletters on AI/ML/DevOps topics from k8s to Agentic Workflows!"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize-phoenix>=13.20.0",
    "boto3>=1.42.61",
    "bs4>=0.0.2",
    "feedparser>=6.0.12",
    "ollama>=0.6.1",
    "openinference-instrumentation>=0.1.46",
    "opentelemetry-exporter-otlp-proto-http>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "psycopg[binary,pool]>=3.3.3",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-json-logger>=4.0.0",
    "requests>=2.32.5",
]
